feat: modularize the earnings processing pipeline
