+++ rust/Cargo.toml
@@ -27,8 +27,6 @@ serde_json = "1"
 miette = "7.6.0"
 colored = "3.1.1"
 tempfile = "3.27.0"
-reqwest = { version = "0.13.2", features = ["json", "rustls"] }
-anyhow = "1.0.102"
 tokio = { version = "1.50.0", features = ["rt-multi-thread", "macros", "time"] }
 tokio-stream = "0.1"
 async-trait = "0.1"
@@ -42,10 +40,11 @@ dotenvy = "0.15.7"
 mongodb = { version = "3.5.1", features = ["bson-3"] }
 textwrap = "0.16.2"
 terminal_size = "0.4.3"
-config = "0.15.22"
 futures = "0.3.32"
 tracing = "0.1.44"
 tracing-subscriber = { version = "0.3.23", features = ["env-filter"] }
+indicatif = "0.18.4"
+console = "0.16.3"
 
 [profile.dev]
 opt-level = 0
