+++ rust/crates/core/src/lib.rs
@@ -1,18 +1,18 @@
-use miette::Diagnostic;
-use thiserror::Error;
-
 pub mod db;
 pub mod domain;
-pub mod earnings_processor;
 pub mod stt;
 
-#[derive(Debug, Error, Diagnostic)]
-pub enum AppError {
-    #[error(transparent)]
-    #[diagnostic(transparent)]
-    Db(#[from] db::DbError),
+pub mod earnings {
+    //! Public API for the earnings pipeline.
 
-    #[error(transparent)]
-    #[diagnostic(transparent)]
-    Stt(#[from] stt::SttError),
+    // Private implementation modules
+    mod errors;
+    mod events;
+    mod processor;
+    mod utils;
+
+    // Public surface
+    pub use errors::{EarningsError, IngestError};
+    pub use events::{EarningsEvent, EarningsObserver, NullObserver};
+    pub use processor::{EarningsProcessor, ProcessRequest};
 }
